refactored application controller to include concern for source feature
