+++ app/controllers/application_controller.rb
@@ -1,14 +1,5 @@
 class ApplicationController < ActionController::Base
   #protect_from_forgery with: exception
   include DeviseWhitelist
-
-  before_action :set_source
-
-  def set_source
-    if params[:q]
-      session[:source] = params[:q]
-    else
-      session[:source] = nil
-    end
-  end
+  include SetSource
 end
